Delete Warehouse: Check Purchase Order closes #187

+++ app/Http/Controllers/WarehouseController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Warehouse;
+use App\Models\PurchaseOrder;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use \App\Http\Requests\StoreWarehouseRequest;
@@ -180,6 +181,14 @@ class WarehouseController extends Controller
     {
         $user = Auth::user();
         if ($user->can('delete warehouses')){
+            /*
+                check if warehouse present in purchase orders
+            */
+            $count = PurchaseOrder::where('warehouse_id', $id)->count();
+
+            if ($count > 0)
+                return redirect(route('warehouses'))->with('error', trans('error.warehouse_has_purchase_order'));
+
             Warehouse::destroy($id);
             return redirect(route('warehouses'))->with('success', trans('app.record_deleted', ['field' => 'warehouse']));
         }
